# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Boilerplate",
        "title": "Go Fiber Boilerplate",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/generic": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "This route needs token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "generic"
                ],
                "summary": "Need token"
            }
        },
        "/users/register": {
            "post": {
                "description": "Register a brand new user to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Details needed for registration",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}